+++ eventmesh-storage/eventmesh-storage-kafka/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.ConnectorResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-kafka=ConnectorResourceServiceKafkaImpl
+kafka=org.apache.eventmesh.storage.kafka.connector.ConnectorResourceServiceKafkaImpl
