+++ eventmesh-storage/eventmesh-storage-kafka/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.producer.Producer
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-kafka=KafkaProducerImpl
+kafka=org.apache.eventmesh.storage.kafka.producer.KafkaProducerImpl
