+++ eventmesh-storage/eventmesh-storage-knative/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.ConnectorResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-knative=ConnectorResourceServiceKnativeImpl
+knative=org.apache.eventmesh.storage.knative.connector.ConnectorResourceServiceKnativeImpl
