+++ eventmesh-storage/eventmesh-storage-pravega/src/main/java/org/apache/eventmesh/storage/pravega/PravegaConnectorResourceServiceImpl.java
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.connector.pravega;
+package org.apache.eventmesh.storage.pravega;
 
 import org.apache.eventmesh.api.storage.ConnectorResourceService;
-import org.apache.eventmesh.connector.pravega.client.PravegaClient;
+import org.apache.eventmesh.storage.pravega.client.PravegaClient;
 
 import lombok.extern.slf4j.Slf4j;
