+++ eventmesh-storage/eventmesh-storage-pravega/src/main/java/org/apache/eventmesh/storage/pravega/PravegaProducerImpl.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.connector.pravega;
+package org.apache.eventmesh.storage.pravega;
 
 import org.apache.eventmesh.api.RequestReplyCallback;
 import org.apache.eventmesh.api.SendCallback;
@@ -24,9 +24,9 @@ import org.apache.eventmesh.api.exception.ConnectorRuntimeException;
 import org.apache.eventmesh.api.exception.OnExceptionContext;
 import org.apache.eventmesh.api.producer.Producer;
 import org.apache.eventmesh.common.config.Config;
-import org.apache.eventmesh.connector.pravega.client.PravegaClient;
-import org.apache.eventmesh.connector.pravega.config.PravegaConnectorConfig;
-import org.apache.eventmesh.connector.pravega.exception.PravegaConnectorException;
+import org.apache.eventmesh.storage.pravega.client.PravegaClient;
+import org.apache.eventmesh.storage.pravega.config.PravegaConnectorConfig;
+import org.apache.eventmesh.storage.pravega.exception.PravegaConnectorException;
 
 import java.util.Properties;
 import java.util.concurrent.atomic.AtomicBoolean;
