+++ eventmesh-storage/eventmesh-storage-pravega/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.connector.ConnectorResourceService
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-pravega=org.apache.eventmesh.connector.pravega.PravegaConnectorResourceServiceImpl
+pravega=org.apache.eventmesh.storage.pravega.PravegaConnectorResourceServiceImpl
