+++ eventmesh-storage/eventmesh-storage-pravega/src/main/resources/META-INF/eventmesh/org.apache.eventmesh.api.producer.Producer
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-pravega=org.apache.eventmesh.connector.pravega.PravegaProducerImpl
+pravega=org.apache.eventmesh.storage.pravega.PravegaProducerImpl
